+++ apps/app/components/static/content/Main.vue
@@ -8,9 +8,15 @@
   -->
 
 <script setup lang="ts">
-import type AppConfig from "~/app.config"
+import { useImagePreview } from "~/composables/useImagePreview"
+import ImagePreview from "~/components/common/ImagePreview.vue"
 
-const props = defineProps<{ post: any, setting: typeof AppConfig }>()
+const props = defineProps<{
+  post: any
+  setting?: any
+}>()
+
+const { images, previewRef } = useImagePreview()
 
 // 正文
 const editorDom = props.post.editorDom?.replaceAll("contenteditable=\"true\"", "contenteditable=\"false\"") ?? ""
@@ -43,6 +49,8 @@ const VNode = () =>
         v-embedblock
         v-fold
         v-desc
+        v-htmlblock
+        v-echarts
         class="protyle-wysiwyg protyle-wysiwyg--attr"
         spellcheck="false"
         contenteditable="false"
@@ -55,6 +63,9 @@ const VNode = () =>
         <VNode />
       </div>
     </div>
+    <client-only>
+      <ImagePreview ref="previewRef" :images="images as any" />
+    </client-only>
   </div>
 </template>