+++ apps/app/plugins/018.echarts.server.ts
@@ -1,0 +1,19 @@
+/*
+ *            GNU GENERAL PUBLIC LICENSE
+ *               Version 3, 29 June 2007
+ *
+ *  Copyright (C) 2024 Terwer, Inc. <https://terwer.space/>
+ *  Everyone is permitted to copy and distribute verbatim copies
+ *  of this license document, but changing it is not allowed.
+ */
+
+/**
+ * ECharts 渲染插件 - 服务端
+ *
+ * @author terwer
+ * @version 1.0.0
+ * @since 0.0.1
+ */
+export default defineNuxtPlugin(({ vueApp }) => {
+  vueApp.directive("echarts", {})
+})