+++ build-node-provider.sh
@@ -1,4 +1,6 @@
 #!/bin/sh
 
+rm -rf dist
+mkdir -p dist/node
 pnpm build -F @terwer/share-pro-app --force -- --from node
 echo "Build node provider finished."